Task2_Create_API_Endpoints_for_CRUD_Operations/api_mongo/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.1.1
pymongo==4.3.3
pydantic==2.6.1
python-dotenv==1.0.1 